+++ openfisca_france/model/prelevements_obligatoires/impot_revenu/ir.py
@@ -1394,13 +1394,13 @@ class microentreprise(Variable):
 class plus_values(DatedVariable):
     column = FloatCol
     entity_class = FoyersFiscaux
-    label = u"Plus_values"
+    label = u"Taxation des plus_values"
 
     @dated_function(start = date(2007, 1, 1), stop = date(2007, 12, 31))
     def function_20070101_20071231(self, simulation, period):  # f3sd is in f3vd holder
         """
-        Taxation des plus value
-        TODO: f3vt, 2013 f3Vg au barème / tout refaire
+        Taxation des plus values
+        TODO: 2013 f3Vg au barème / tout refaire
         """
         period = period.this_year
         f3vg = simulation.calculate('f3vg', period)
@@ -1437,7 +1437,7 @@ class plus_values(DatedVariable):
     def function_20080101_20111231(self, simulation, period):  # f3sd is in f3vd holder
         """
         Taxation des plus value
-        TODO: f3vt, 2013 f3Vg au barème / tout refaire
+        TODO:  2013 f3Vg au barème / tout refaire
         """
         period = period.this_year
         f3vg = simulation.calculate('f3vg', period)
@@ -1476,12 +1476,13 @@ class plus_values(DatedVariable):
     def function_20120101_20121231(self, simulation, period):  # f3sd is in f3vd holder
         """
         Taxation des plus value
-        TODO: f3vt, 2013 f3Vg au barème / tout refaire
+        TODO: 2013 f3Vg au barème / tout refaire
         """
         period = period.this_year
         f3vg = simulation.calculate('f3vg', period)
         f3vh = simulation.calculate('f3vh', period)
         f3vl = simulation.calculate('f3vl', period)
+        f3vt = simulation.calculate('f3vt', period)
         f3vm = simulation.calculate('f3vm', period)
         f3vi_holder = simulation.compute('f3vi', period)
         f3vf_holder = simulation.compute('f3vf', period)
@@ -1519,13 +1520,14 @@ class plus_values(DatedVariable):
     def function_20130101_20151231(self, simulation, period):  # f3sd is in f3vd holder
         """
         Taxation des plus value
-        TODO: f3vt, 2013 f3Vg au barème / tout refaire
+        TODO: 2013 f3Vg au barème / tout refaire
         """
         period = period.this_year
         f3vg = simulation.calculate('f3vg', period)
         f3vh = simulation.calculate('f3vh', period)
         f3vl = simulation.calculate('f3vl', period)
         f3vm = simulation.calculate('f3vm', period)
+        f3vt = simulation.calculate('f3vt', period)
         f3vi_holder = simulation.compute('f3vi', period)
         f3vf_holder = simulation.compute('f3vf', period)
         f3vd_holder = simulation.compute('f3vd', period)