+++ CHANGELOG.md
@@ -14,6 +14,31 @@
 	* `place_hebergement`
 	* `ada`
 
+## 4.1.18
+
+* Introduce PCH
+
+## 4.1.17
+
+* Reimplement `rsa_indemnites_journalieres_activite` and `date_arret_de_travail` using `datetime.date.min`.
+
+## 4.1.16
+
+* Correct wrong behaviour in the combination of exemptions with the dispositif Jeune Entreprise Innovante (JEI)
+
+
+## 4.1.15
+
+* Consider some indemnités journalieres as revenus de remplacement in PPA (and RSA), according to the date of the arret de travail.
+* Introduce
+	* `rsa_indemnites_journalieres_hors_activite`
+	* `rsa_indemnites_journalieres_activite`
+	* `date_arret_de_travail`
+
+## 4.1.14
+
+* Add cerfa boxe : 3vt (PEA)
+
 ## 4.1.13
 
 * Increase RSA base amount to 535.17 from September 2016